Harmoize usage of async fn when deal with HTTP and networking code
Previously, we used a mix of `Runtime::block_on` and `Runtime::spawn`
to kick off all the important background worker processes that make
cnd work.

This patch harmonizes this usage by:

- Always using `Runtime::spawn` to spawn background workers
- Always using `Runtime::block_on` to execute code that should
run synchronously as we start up cnd.

To achieve this, we separate binding to the socket for the HTTP API
(something that should happen synchronously during startup because
we want to terminate if we fail) from the actual worker future
that handles the incoming requests.

We also separate creating an instance of `Swarm` from actually
spawning the background worker to align it with the HTTP API
background worker.

This introduces the `std::thread::park` function call again to idle
the main process.

This patch also prepares for potential follow-up work where we want
to gracefully shutdown by listening on signals like SIGINT. Such code
would go in place of the current `std::thread::park`.