Duplicate the state traits
Currently we have three traits, `Update`, `Get`, and `Insert` that are
used to manage the `LedgerStates`.  These traits and their associated
methods use `SwapId` as the key to the state stores (hash maps).  We
just added two new swap identifiers, `LocalSwapId` and `SharedSwapId`
for use with the new 'spilt protocols' (hand, hecr20 etc.).  We would
like to have the ledger state store functionality available for usage
with the split protocols but we want to keep rfc003 code functioning.

Copy the ledger state code into the rfc003 module.  Replace `SwapId`
with `LocalSwapId` in the original ledger state code.

Note: The main network behaviour now contains multiple ledger states, a
pair for the rfc003 protocol and a bunch for the spilt protocols.