Do not panic during halight communication if no lnd
In the main, only the path to lnd certificate and macaroon
were passed. The files were read only during the finalize
step of the communication, leading to a panic if not
present or unreadable.

The logic has been modified to provide the following
functionality:
- An attempt is made to read the files when starting cnd,
the result is an `Option` passed down to the swarm.
- When a halight swap is created on the REST API,
the `Option` is checked and an error is returned to the
caller if it's `None`
- During the finalize phase, we assume that the check was
previously done so we should not be able to finalize a
halight swap if the `Option` is `None`. However, if there
is a bug, we log an error a do not finalize (nor panic).