Avoid double download of lnd binary
When investigating the spawn ETXTBSY error I noticed
that both alice and bob try to download the lnd binary.
The spawn error happens for lnd and in the logs we see
alice's download finishing but it never does for bob.

Best case, it fixes the spawn issue, worst case, it
saves time and bandwidth.