Wait for a transaction to be mined before releasing the lock.
Our Geth test node mines a block every second. When sending a
transaction we set the nonce to be the value that we first read from the
blockchain, if we try to do this twice within a block it fails (Geth
thinks we are trying to do a transaction replacement).

Signed-off-by: Philipp Hoenisch <philipp@hoenisch.at>