Save what we learned from the other party to the database
To achieve this, we first need to save the created swap in
the facade of the HTTP controller, otherwise we don't
have a foreign key we can attach our data to.

To save the created swap, we generically implement the Save
trait over any combination of `CreatedXYZ` struct that implements
the necessary conversion traits provided by the database layer.

Co-authored-by: Daniel Karzel <daniel.karzel@coblox.tech>