Rename ComitLN
During the LN squad we name the 'new' network behaviour as
`ComitLN` (the original was, at this time, name `Comit`). Since then the
original comit network behaviour has been renamed to `rfc003_comit`
paving the way for us to rename the new behaviour to `Comit`.

Rename the `ComitLN` network behaviour to be just `Comit` - it is _the_
comit network behaviour (soon as rfc003 disappears).