Move swap_id to root module
The `LocalSwapId` type is not confined to the `swap_protocols` module in
any way, move it up to the root module. While we are at it, `pub use`
it so that users of this type do not need to care what module it is
in, this makes it easier to move code around.