Use types from crate in cnd
Instead of bringing types into scope from the comit crate we can use
them from cnd crate because we now re-export comit types within cnd.
This makes it easier to import types because devs do not need to
remember what is in the comit crate and what is in cnd. It also makes
it easier to move code between the two crates as needed.