Move time-related methods from cache to dedicated modules
The `Cache` components are designed to be proxies. Hence, they
must not expose any APIs other than the ones from the component
they are proxying.

We achieve that by moving the calculation of Bitcoin's median-time
to the `bitcoin` module in the `comit` crate.

For Ethereum, the impl is so trivial that we just do it inline in
the facade.