Rename halight to halbit
The protocol name halgiht is not adequate to fully describe the
protocol because it does not stipulate the asset i.e., the layer one
blockchain involved.

The name halbit (Htlc Asset Lightning - Bitcoin) fully describes the
protocol. We use 'Asset' just to make the acronym easier to read and
pronounce.

Re-name halight to halbit across the whole codebase.