Reduce noise in imports
The import statements in the storage module are too noisy.

The storage module fully 'owns' the database, it makes sense to just
bring all the types from the db into scope.

In an effort to move the whole code base in this direction we can also
export `Get` up to the root level module and import it without the
qualified path.