Implement connected_peers on the Facade
In order to move away from the `rfc003Facade` we will need the
`Facade` to be able to provide the functionality for getting a list of
connected peers. Currently, in the 'new' code peer connections are
handled low down in the comit crate.

Expose `connected_peers` from the `Comit` network behaviour. Implement
`ComitPeers` on the `Facade` calling through to the newly exposed
functionality.