Service / endpoint with the Facade
In order to move away from using the `Rfc003Facade` we need to be able
to service requests to the `/` endpoint (incl. siren).

Implement `ListenAddresses` on the `Facade`. Use the `Facdade` when
servicing GET requests to `/` via the `get_info()` and
`get_siren_info()` functions.

Since we now handle the `/` endpoint using 'new' code do not check for
`/swaps/rfc003` when testing the siren response for requests to `/`.