e2e: Remove wrong peer id test
This test is based on rfc003 code. For the new swap protocols we have
yet to decide how we are going to handle peer ids (and ip addresses).
It is not an error in the current code to announce a swap to a random
peer id. This test is not useful for the new swap protocols.

In preparation for removing rfc003 remove the wrong peer id test.