e2e: Remove multiple peers tests
We currently have a single test in `peers_using_ip.ts`, this test is
actually a 'multiple peers' test.

The current test framework is not written to correctly handle multiple
peers using the new swap protocols. The Actor class make a bunch of
assumptions including such things as

- there is only a singe swap
- ledgers/wallets are initialized when we create a swap
- actor name is linked to the cryptographic protocol role

For theses reasons the tests that use `threeActorTest` use rfc003
actors and rfc003 swaps.

We do want to be able to test swaps to multiple peers, to do so with
the new swap protocols requires heavy modification of current test
infrastructure.

In preparation for removing rfc003 elect to delete the rfc003 multiple
peer tests.