e2e: Remove the threeActorTest
The `threeActorTest` does not function correctly with the new swap
protocols. We recently removed all the rfc003 tests that use this
function.

Having a third actor like we do for one actor and two actors is
misleading because swap protocols only involve two parties. Actor
names are intrinsically tied to the cryptographic role the actor plays
so having a third actor implies some different cryptographic role -
this is not the case. Actually the third party is just _another_ actor
in the role of ether Alice or Bob doing a separate swap with one of
the initial actors. We are mixing up cryptographic roles with the
concept of a 'user' using a cnd node to do a swap.

Remove the `threeActorTest` because it is confusing and not useful for
the new swap protocols.