Implement Hbit Redeem using Execute trait
No need for using the `Do` trait, since there is no decision-making
when it comes to Bob's redeem: he should always do it, and once it's
done (or Alice beats him by refunding) the protocol ends.

But, we introduce this to-do:

TODO: Prevent Bob from trying to redeem again (applies to the
all the refunds too). Reusing the Do trait seems wrong since we
should never abort at this stage, which is why we used the
Execute trait directly. There is no risk in doing this action
more than once, but it's a bit wasteful. We should probably
introduce another trait which composes CheckMemory, Execute and
Remember to solve this problem (P.S. naming is hard).