Revert making WatchOnly impls no-ops
This was motivated by Nectar needing to free funds once a swap had
finished. Since we couldn't possibly know if the taker would ever
actually refund, we could get stuck waiting for them to refund,
keeping the funds as reserved even though the swap would never go to
completion.

Unfortunately, this needs to be reverted as a no-op resolves
immediately, meaning that we never give ourselves a chance to redeem
Beta (there is a race between the taker refunding and Nectar
redeeming).

The design of the swap execution functions can be improved in the
future, but for the time being we can assume that the taker's
motivation to refund will prevail and it will always happen.
Furthermore, in the cases where Nectar actually gets to refund, we
could send an event up to the "main loop" to free the funds
immediately, without needing to wait for the taker to refund
themselves.