Remove announce trading pair
All the announce trading pair logic is superfluous. We only support a
single trading pair (this implies only a single gossipsub topic). We
subscribed to this topic when the orderbook is created. There is no
need for users of cnd to be able to announce a trading pair.

Remove all the code associated with announcing a trading pair from
comit and cnd.