cnd: Remove herc20 hbit from orderbook identifiers
We only support a single order type. The order does not map to the
COMIT protocol directly but rather is dependent on the order
`Position`. It is misleading to use herc20 and hbit in the identifiers
for the orderbook because it is not how we use the terms throughout
the rest of cnd.