Unconditionally reserve funds before respawning
Given that we only free funds once the swap execution is over, we need
to make sure that we always reserve funds before respawning (since the
`Maker`'s state is lost between restarts of Nectar).

Additionally, we could examine swaps and decide whether we have enough
money or not to continue with the execution successfully, but it would
be quite complex. Instead, we let the swap execution fail, which is
functionally equivalent to preventing the respawn.