Upgrade to latest version of COMIT lib
This allows us to support both buy and sell orders!

It also allows us to set expiries properly when creating offers.

We add missing implementations to enable calling `nectar_herc20_hbit`.

The main limitation now is that orders are limited to a local
development network. This should be easy to fix once we include the
network as part of the `dai::Asset` and a possible
`nectar::bitcoin::Asset`.