Save address hint in the peer tracker
The peer tracker is our component that takes care of peer and address
discovery.

Currently when the `addresses_of_peers()` method is called we return a
list of the addresses of peers we are connected to, this is not that
useful if we are trying to find out an address to dial to make the
connection in the first place.

Add a map of peer ids to address hints and return this hint if one
exists when the `addresses_of_peer()` method is called.