Properly connect two actors with each other
Previously, we used the OrderFactory to connect two actors and only
waited for a single second for that two second.

Firstly, connecting two actors is unrelated to the concern of the
factory design pattern, hence this function moves to the actor class.
Secondly, waiting for a single second is error-prone because it
depends on the CPU speed of the current machine. Instead, we poll
cnd until it reports a connection to the given peer.