Add Standardized COMIT Expiry Times module
- Add a module `expiries`.

- Implement timestamp improvements

We would like to be able to manipulate timestamps and durations
together. Implement a bunch of improvements to our timestamp module to
facilitate this. Use the improvements in `expiries` module.

Notes: We use `time::Duration` because it is more suited to durations
of the timeframes we expect i.e., hour long durations (`std::time` is
better suited, according to the docs, to system times i.e., smaller
timeframes).