Add a CLI argument to specify the COMIT network one wants to connect to
Using the given network, the user can control the defaults within the
config file if the configuration is not present.

If there is a configuration present, it has to match what the default
would have generated. This avoids situations in which the user says:

> cnd --network main

But the config file says:

[bitcoin]
network = testnet