Add a scaling factor
When calculating whether a counterparty will continue with the swap we
use time to mine N blocks as a heuristic. By definition this depends
on N but the actual time to mine N blocks may vary wildly.

If we use the same N to calculate the expiry offsets that we use
during the swap to determine if the counterparty will continue we can
easily cause a swap to be aborted by a longer than usual block time.

In order to counteract this we add a scaling factor, we use the
scaling factor to scale the expiry offsets. This of course has an
effect on the option problem but hopefully this is apparent to users.