Implement new swap response format
This format favours a flat design in which events are appended
to a list as they are observed by cnd. Or that is at least what
we should do. With the current implementation in cnd, this is not
possible because we compute this state every time the user asks for
it.

In the long run however, we can enrich this format with timestamps
and other information and actually save those events to the database
so that we just need to read them.

This also radically simplifies the assertions in our test.