Remove PATCH semvar number from all libraries
In semvar changing the PATCH version by definition should not break
the usage of the library.

In Nectar we depend on comit, if we use explicit patch versions for a
dependency that we have in common with cnd then the patch versions
have to match. This adds unnecessary maintenance burden on the two
repositories.

Remove the patch version number from all dependencies.