Switch to Price/Quantity model
The goal of this commit is bring nectar up to date with comit's order model and reduce type duplication. Changed the BtcDaiOrderForm struct to use the price quantity model that was recently adopted by comit-rs. The bitcoin and dai asset types were removed because the btc and eth networks were no longer required. The PRECISION constant was fixed which was causing a 1 decimal place conversion error".