Introduce a dedicated BalanceAsserter
This abstractions replaces our previous way of asserting balances
after a swap. The advantage we hope to get out of this is to make
it more explicit how the assertion works for specific assets. For
example, we can now be explicit about the fact fees only matter
when dealing with Bitcoin on-chain.