Introduce BitcoinFaucet
This fixes a tricky bug where we called `pollUntilMinted` a 2nd time
when trying to mint money to the LN wallet but checked the balance
of the actor's BitcoinWallet.

To mitigate this, we separate out the code for minting into a dedicated
faucet that also encapsulates the whole getStartingBalance + pollUntilMinted
dance.