Segregate trading strategy
The current trading strategy that takes for input:
- a spread
- mid-market rate
- available balances

Is now in a dedicated module. Allowing this concern to be removed from
the maker.
Some primitives such as the bitcoin fees and the reserved assets for
ongoing swaps are also moved in the strategy module. It could be argued
that they should stay with the maker or moved in yet another module.

I propose that such concern can be reviewed when another trading
strategy is implemented, validating the best scope for trading
strategies.