Pre-build binaries for e2e tests
Our e2e tests lazily build the cnd and nectar binaries with
`cargo build -p`. That is different from running `cargo build --all-targets`
as the latter includes `cfg(test)`. As such, the e2e tests time out
because the CI machines are not fast enough to actually build the
binaries before the tests timeout.

We add a dedicated step to the CI workflow the builds the two binaries.