Remove chrono dependency in favor of `time`
We were using two different time libraries before. This patch unifies
the codebase to only use `time:0.2`.

`diesel` has a feature-flag for interoperability with `chrono` that we
were using. Hence we need to now define our own serialization. Contrary
to what the feature-flag did, we are now storing time as a unix timestamp.
This means we have to run a migrate timestamps in existing databases to
the new format.