Introduce db::commands module
This is equivalent to db::queries but performs updates to the DB
instead of just reading from it. We borrow the name from the CQRS
pattern.

The CQRS pattern is helpful in cases where the domain doesn't align
well with CRUD due to differences in the patterns of how data is
read vs written. It allows for more flexibility of both cases without
compromise but comes at the cost of some duplication between the two
modules.

However, the APIs of diesel are pretty powerful and compose nicely
as long as one doesn't want to reuse too much code between different
functions. Splitting things into queries and commands fits nicely here
because we can implement both completely on top of diesel's APIs
without worrying about code reuse too much.