Optimize btsieve logs
The current design leads to several lines per block in the log file.
This creates a visual noise that is hard to follow.

With this patch, we pack more information into the context (like the
number of transactions) and only ever emit a single event during the
matching process.

We also remove all events from the connector and the cache as those
did not add any value.

Finally, we change the span log levels from trace to error. Span log
levels are tricky to wrap your head around because they work a bit
counter-intuitively. Basically, a span log level says up to which
event level the span should be included. In other words, a trace-span
will not show up for debug events but an error-span would. Our spans
contain the block and tx hash and should hence always be included.