Improve max possible fee representation
This fee is only needed for Bitcoin at this stage
so changed it to BTC specific.

It is unrealistic to ask the user to know the max fee for a COMIT tx so
instead split the settings in 2 value: sat/vbyte fee and tx size.

Tx size is set to a safe bet. It can of course vary depending of the
number of inputs used by the user. Hence it is configurable.