Downgrade Rust to 1.45
1.46 introduced a regression during codegen that made the
type-length-limit and thus the compile time increase expontially
for nested async functions.

We downgrade to Rust 1.45 for the time being to fix this.
As a result, we have to make some functions non-const again. However,
that is not a big deal as we don't really benefit from their constness
at the moment anyway. Once we `time` is upgraded to support more
`const` functions, this will get more interesting.

Reference: https://github.com/rust-lang/rust/issues/75992