Ensure Bob can complete swap if Alice is slow
Once a swap has started, Bob should not take in account the full
"alice time to act" duration to know whether it can
complete the swap. This duration is only useful to know the total time
needed for happy path when calculating the expiries.

Also improve the values used in the test so they are less arbitrary.

Finally, add a test to confirm that Alice is told not
to redeem if she cannot secure her redeem transaction
before beta expiry (bob refund) is reached.